% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R
\name{clean_up}
\alias{clean_up}
\title{Clean up before reprocessing data files}
\usage{
clean_up()
}
\description{
Cleans the data folder of all entries. Cleans R of all files except those
for that are data R files. Cleans up the pkgdown docs folder also.
}
\keyword{internal}
