% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Columbia-River-Chum-Salmon-ESU.R
\docType{data}
\name{Columbia-River-Chum-Salmon-ESU}
\alias{Columbia-River-Chum-Salmon-ESU}
\title{Columbia River Chum Salmon ESU}
\format{
A csv file.
\describe{
\item{NMFS_POPID}{NWFSC id for the population.}
\item{ESU}{}
\item{SPECIES}{}
\item{RUN_TIMING}{}
\item{MAJOR_POPULATION_GROUP}{}
\item{COMMON_POPULATION_NAME}{}
\item{BROOD_YEAR}{The year that the count was made.}
\item{NUMBER_OF_SPAWNERS}{Spawners count. -99 means missing count.}
\item{FRACWILD}{Fraction wild. -99 means missing information.}
}
}
\description{
Spawner counts and fraction wild data.
}
\details{
Downloaded from the SPS Master Oracle 2013-2014 Database on
January 28, 2016
}
\examples{
data('Columbia-River-Chum-Salmon-ESU')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=NUMBER_OF_SPAWNERS)) + geom_point(na.rm = TRUE) +
  ggtitle('Columbia-River-Chum-Salmon-ESU') +
  facet_wrap(~COMMON_POPULATION_NAME)
}
