% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vignette.R
\name{create_vignette}
\alias{create_vignette}
\title{Create vignette}
\usage{
create_vignette(dataname, yearname)
}
\arguments{
\item{dataname}{Name of the data}

\item{yearname}{YEAR or BROOD_YEAR}
}
\description{
Create the vignette for the data
}
\keyword{internal}
