% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_text.R
\name{describe_text}
\alias{describe_text}
\title{describe_text}
\usage{
describe_text(thecols)
}
\arguments{
\item{thecols}{the column names from the file}
}
\description{
Adds the format code for the Roxygen data header by reading in the
columns names and only adding the needed items.
}
\keyword{internal}
