% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hood-Canal-HCchum2020.R
\docType{data}
\name{Hood-Canal-HCchum2020}
\alias{Hood-Canal-HCchum2020}
\title{Hood-Canal-HCchum2020}
\format{
A csv file.
\describe{
\item{SERIES}{For certain populations, there are multiple types of spawner count data. For example survey and model counts that are based on different methods. In this case, the 'Series' column gives a name for the dataset to distinguish it from other data sets of the same population.}
\item{ESU}{ESU or DPS name.}
\item{SPECIES}{Species name.}
\item{METHOD}{Method for the spawner count. Survey refers to a spawning ground survey; the count itself is an expansion. See the SPS (database link in sources below) for the details for the data. LadderCount is a weir or dam count. Model  refers to the DABOM model that expands from PIT tag data. GSI is an expansion based on the BOXCAR model, which uses GSI to allocate fish sampled at Lower Granite Dam.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{POPULATION_NAME}{Long population name.}
\item{COMMON_POPULATION_NAME}{Shorter population name used in Viability Report figures and tables.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
\item{CONTRIBUTOR}{Data contributor. Listed under Source below.}
\item{COMPILER}{NWFSC staff who compiled the data file: Mari Williams}
\item{CITATION}{Data citation. Listed under References below.}
}
}
\source{
WDFW, Point No Point Treaty Council
}
\description{
Spawner and fraction wild data. Species: Summer Chum. Method: . \href{../data/Hood-Canal-HCchum2020.html}{View raw data} or \href{../data/Hood-Canal-HCchum2020.csv}{Download raw data}

Spawner and fraction wild data. Species: Summer Chum.
\itemize{
\item Strait of Juan de Fuca: Survey. \href{../data/Hood-Canal-HCchum2020.html}{View raw data} or \href{../data/Hood-Canal-HCchum2020.csv}{Download raw data}
}

Spawner and fraction wild data. Species: Summer Chum.
\itemize{
\item Strait of Juan de Fuca: . \href{../data/Hood-Canal-HCchum2020.html}{View raw data} or \href{../data/Hood-Canal-HCchum2020.csv}{Download raw data}
}

Spawner and fraction wild data. Species: Summer Chum.
\itemize{
\item Hood Canal: Survey. \href{../data/Hood-Canal-HCchum2020.html}{View raw data} or \href{../data/Hood-Canal-HCchum2020.csv}{Download raw data}
}
}
\examples{
data('Hood-Canal-HCchum2020')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=YEAR, y=NUMBER_OF_SPAWNERS)) + 
  geom_point(na.rm = TRUE) +
  ggtitle('Hood-Canal-HCchum2020') +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=YEAR, y=FRACWILD)) +
  geom_point(na.rm = TRUE) +
  ggtitle('Fraction Wild') +
  ylim(0,1) +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))
}
\references{
Ford, M.J., et al. 2022. Biological Viability Assessment Update for Pacific Salmon and Steelhead Listed Under the Endangered Species Act: Pacific Northwest. U.S. Department of Commerce, NOAA Technical Memorandum NMFS-NWFSC-171. https://doi.org/10.25923/kq2n-ke70

Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}

Original source:
\itemize{
\item Strait of Juan de Fuca: Hood Canal Summer Chum Natural Origin Spawner Abundance (1971-2019). Spawner abundance data. Point No Point Treaty Council and Washington Department of Fish and Wildlife. Personal Communication with Mindy Rowse, 2020.
\item Hood Canal: Hood Canal Summer Chum Natural Origin Spawner Abundance (1968-2019). Spawner abundance data. Point No Point Treaty Council and Washington Department of Fish and Wildlife. Personal Communication with Mindy Rowse, 2020.
}
}
