% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_raw.R
\name{process_data_raw}
\alias{process_data_raw}
\title{Process data-raw}
\usage{
process_data_raw()
}
\description{
Process a data file in data-raw and create an documentation R file
in the R directory with an Roxygen header. This also creates files
in docs/data so that user can download the raw data.
}
\details{
\enumerate{
\item Delete all files in R dir except the R files need to process data files
\item Delete all data in data dir
\item Delete all vignettes in vignettes dir
}
}
\keyword{internal}
