% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Purple-Fish-River.R
\docType{data}
\name{Purple-Fish-River}
\alias{Purple-Fish-River}
\title{Purple Fish River}
\format{
A csv file.
\describe{
\item{BROOD_YEAR}{The year that the redd count was made.}
\item{SPAWNERS}{Spawners estimated from the redd counts. See details.}
\item{FRACWILD}{Fracwild wild. This is assumed to be 1 for this population due to the lack of any nearby hatcheries.}
}
}
\description{
Redd counts from upper Purple Fish Creek.
}
\details{
Here are all the details. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi at sem massa. Phasellus ac nulla metus. Proin aliquet nisl libero, a varius urna mattis vitae. Fusce a sapien iaculis, aliquam mauris ac, maximus risus. Donec placerat leo a dolor accumsan laoreet. Vivamus cursus, mi id efficitur fringilla, massa enim semper lacus, posuere sagittis ante purus id nibh. Aenean at enim iaculis, elementum risus ut, tempus dui. Pellentesque vel tincidunt nisi, non suscipit turpis. Phasellus finibus non orci vel faucibus. Nullam placerat tellus et ipsum sagittis sollicitudin. Nullam accumsan, lorem et dictum suscipit, ante risus porta nisl, et volutpat mi mauris eleifend elit. Sed tristique finibus placerat. Integer at luctus dui, at imperdiet odio.
}
\examples{
data('Purple-Fish-River')
library(ggplot2)
ggplot(out, aes(x=BROOD_YEAR, y=SPAWNERS)) + geom_point(na.rm = TRUE) +
  ggtitle('Purple-Fish-River')
}
\author{
{Jenny John}
}
