% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Red-Fairy-River.R
\docType{data}
\name{Red-Fairy-River}
\alias{Red-Fairy-River}
\title{Red Fairy River}
\format{
A csv file.
\describe{
\item{BROOD_YEAR}{The year that the redd count was made.}
\item{SPAWNERS}{Spawners estimated from the redd counts. See details.}
\item{FRACWILD}{Fracwild wild. This is assumed to be 1 for this population due to the lack of any nearby hatcheries.}
}
}
\description{
Redd counts taken at the weir below Green Fairy Creek.
}
\details{
Here are all the details.
}
\author{
{Jenny J.}
}
