% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Salmon Chinook (Snake River spring-summer-run
%   ESU)-ICSRSSchinookModel.R
\docType{data}
\name{Salmon Chinook (Snake River spring-summer-run ESU)-ICSRSSchinookModel}
\alias{Salmon Chinook (Snake River spring-summer-run ESU)-ICSRSSchinookModel}
\title{Salmon Chinook (Snake River spring-summer-run ESU)-ICSRSSchinookModel}
\format{
A csv file.
\describe{
\item{Series}{For certain populations, there are multiple types of spawner count data. For example survey and model counts that are based on different methods. In this case, the 'Series' column gives a name for the dataset to distinguish it from other data sets of the same population.}
\item{Method}{Method for the spawner count. Survey refers to a spawning ground survey of some type. See the SPS (database link in sources below) for the details for the data. LadderCount is a weir or dam count.}
\item{Citation}{Data citation.}
\item{Contributor}{Data contributor.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{ESU}{ESU or DPS name.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{COMMON_POPULATION_NAME}{Population name.}
\item{SPECIES}{Species name.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{BROOD_YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
}
}
\source{
Nez Perce Tribe
}
\description{
Spawner and fraction wild data. Species: Chinook salmon. Method: Model
}
\examples{
data('Salmon Chinook (Snake River spring-summer-run ESU)-ICSRSSchinookModel')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=NUMBER_OF_SPAWNERS)) + 
  geom_point(na.rm = TRUE) +
  ggtitle('Salmon Chinook (Snake River spring-summer-run ESU)-ICSRSSchinookModel') +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), strip.text.x = element_text(size = 6))

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=FRACWILD)) +
  geom_point(na.rm = TRUE) +
  ggtitle('Fraction Wild') +
  ylim(0,1) +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), strip.text.x = element_text(size = 6))
}
\references{
Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}

Original source: Citations with the format CAXMMDDYYYY indicate that the data are from the Coordinated Assessments Partnership (CAP) database and were downloaded on on MM-DD-YYYY. \url{https://www.streamnet.org/home/data-maps/fish-hlis/}.
\itemize{
\item Tucannon River: CAX05262020
\item Tucannon River: CAX05262021
\item Tucannon River: CAX05262022
\item Tucannon River: CAX05262023
\item Tucannon River: CAX05262024
\item Tucannon River: CAX05262025
\item Tucannon River: CAX05262026
\item Tucannon River: CAX05262027
\item Lookingglass Creek: CAX05262029
\item Lookingglass Creek: CAX05262030
\item Lookingglass Creek: CAX05262031
\item Lookingglass Creek: CAX05262032
\item Lookingglass Creek: CAX05262033
\item Lookingglass Creek: CAX05262034
\item Lookingglass Creek: CAX05262035
\item Lookingglass Creek: CAX05262036
\item Lookingglass Creek: CAX05262037
\item Lookingglass Creek: CAX05262038
\item Catherine Creek: CAX05262040
\item Catherine Creek: CAX05262041
\item Catherine Creek: CAX05262042
\item Catherine Creek: CAX05262043
\item Catherine Creek: CAX05262044
\item Catherine Creek: CAX05262045
\item Catherine Creek: CAX05262046
\item Catherine Creek: CAX05262047
\item Catherine Creek: CAX05262048
\item Catherine Creek: CAX05262049
\item Grande Ronde River Upper Mainstem: CAX05262050
\item Grande Ronde River Upper Mainstem: CAX05262051
\item Grande Ronde River Upper Mainstem: CAX05262052
\item Grande Ronde River Upper Mainstem: CAX05262053
\item Grande Ronde River Upper Mainstem: CAX05262054
\item Grande Ronde River Upper Mainstem: CAX05262055
\item Grande Ronde River Upper Mainstem: CAX05262056
\item Imnaha River Mainstem: CAX05262057
\item Imnaha River Mainstem: CAX05262058
\item Imnaha River Mainstem: CAX05262059
\item Imnaha River Mainstem: CAX05262060
\item Imnaha River Mainstem: CAX05262061
\item Imnaha River Mainstem: CAX05262062
\item Imnaha River Mainstem: CAX05262063
\item Imnaha River Mainstem: CAX05262064
\item Imnaha River Mainstem: CAX05262065
\item Big Sheep Creek: CAX05262066
\item Big Sheep Creek: CAX05262067
\item Big Sheep Creek: CAX05262068
\item Big Sheep Creek: CAX05262069
\item Big Sheep Creek: CAX05262070
\item Big Sheep Creek: CAX05262071
\item Big Sheep Creek: CAX05262072
\item Big Sheep Creek: CAX05262073
\item Big Sheep Creek: CAX05262074
\item Little Salmon River: CAX05262075
\item Little Salmon River: CAX05262076
\item Little Salmon River: CAX05262077
\item Little Salmon River: CAX05262078
\item Little Salmon River: CAX05262079
\item Little Salmon River: CAX05262080
\item Secesh River: CAX05262081
\item Secesh River: CAX05262082
\item Secesh River: CAX05262083
\item Secesh River: CAX05262084
\item Secesh River: CAX05262085
\item Secesh River: CAX05262086
\item Secesh River: CAX05262087
\item Secesh River: CAX05262088
\item Secesh River: CAX05262089
\item Secesh River: CAX05262090
\item East Fork South Fork Salmon River: CAX05262091
\item East Fork South Fork Salmon River: CAX05262092
\item East Fork South Fork Salmon River: CAX05262093
\item East Fork South Fork Salmon River: CAX05262094
\item East Fork South Fork Salmon River: CAX05262095
\item East Fork South Fork Salmon River: CAX05262096
\item East Fork South Fork Salmon River: CAX05262097
\item East Fork South Fork Salmon River: CAX05262098
\item East Fork South Fork Salmon River: CAX05262099
\item East Fork South Fork Salmon River: CAX05262100
\item Big Creek: CAX05262101
\item Big Creek: CAX05262102
\item Big Creek: CAX05262103
\item Big Creek: CAX05262104
\item Big Creek: CAX05262105
\item Big Creek: CAX05262106
\item Big Creek: CAX05262107
\item Big Creek: CAX05262108
\item Big Creek: CAX05262109
\item Bear Valley Creek: CAX05262110
\item Bear Valley Creek: CAX05262111
\item Bear Valley Creek: CAX05262112
\item Bear Valley Creek: CAX05262113
\item Bear Valley Creek: CAX05262114
\item Panther Creek: CAX05262115
\item Panther Creek: CAX05262116
\item North Fork Salmon River: CAX05262117
\item North Fork Salmon River: CAX05262118
\item North Fork Salmon River: CAX05262119
\item North Fork Salmon River: CAX05262120
\item Lemhi River: CAX05262121
\item Lemhi River: CAX05262122
\item Lemhi River: CAX05262123
\item Lemhi River: CAX05262124
\item Lemhi River: CAX05262125
\item Lemhi River: CAX05262126
\item Lemhi River: CAX05262127
\item Lemhi River: CAX05262128
\item Lemhi River: CAX05262129
\item Lemhi River: CAX05262130
\item Salmon River Lower Mainstem: CAX05262131
\item Salmon River Lower Mainstem: CAX05262132
\item Salmon River Lower Mainstem: CAX05262133
\item Salmon River Lower Mainstem: CAX05262134
\item Salmon River Lower Mainstem: CAX05262135
\item Salmon River Lower Mainstem: CAX05262136
\item Salmon River Lower Mainstem: CAX05262137
\item Pahsimeroi River: CAX05262138
\item Pahsimeroi River: CAX05262139
\item Pahsimeroi River: CAX05262140
\item Pahsimeroi River: CAX05262141
\item Pahsimeroi River: CAX05262142
\item Pahsimeroi River: CAX05262143
\item Pahsimeroi River: CAX05262144
\item Pahsimeroi River: CAX05262145
\item East Fork Salmon River: CAX05262146
\item East Fork Salmon River: CAX05262147
\item East Fork Salmon River: CAX05262148
\item East Fork Salmon River: CAX05262149
\item East Fork Salmon River: CAX05262150
\item East Fork Salmon River: CAX05262151
\item East Fork Salmon River: CAX05262152
\item Yankee Fork: CAX05262153
\item Yankee Fork: CAX05262154
\item Yankee Fork: CAX05262155
\item Yankee Fork: CAX05262156
\item Yankee Fork: CAX05262157
\item Yankee Fork: CAX05262158
\item Yankee Fork: CAX05262159
\item Yankee Fork: CAX05262160
\item Salmon River Upper Mainstem: CAX05262161
\item Salmon River Upper Mainstem: CAX05262162
\item Salmon River Upper Mainstem: CAX05262163
\item Salmon River Upper Mainstem: CAX05262164
\item Salmon River Upper Mainstem: CAX05262165
\item Salmon River Upper Mainstem: CAX05262166
\item Salmon River Upper Mainstem: CAX05262167
\item Salmon River Upper Mainstem: CAX05262168
\item Salmon River Upper Mainstem: CAX05262169
\item Salmon River Upper Mainstem: CAX05262170
\item Valley Creek: CAX05262171
\item Valley Creek: CAX05262172
\item Valley Creek: CAX05262173
\item Valley Creek: CAX05262174
\item Valley Creek: CAX05262175
\item Valley Creek: CAX05262176
\item Valley Creek: CAX05262177
\item Valley Creek: CAX05262178
\item Valley Creek: CAX05262179
\item Valley Creek: CAX05262180
}
}
