% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Salmon Chinook (Upper Willamette River
%   ESU)-UWchinook2020.R
\docType{data}
\name{Salmon Chinook (Upper Willamette River ESU)-UWchinook2020}
\alias{Salmon Chinook (Upper Willamette River ESU)-UWchinook2020}
\title{Salmon Chinook (Upper Willamette River ESU)-UWchinook2020}
\format{
A csv file.
\describe{
\item{Series}{For certain populations, there are multiple types of spawner count data. For example survey and model counts that are based on different methods. In this case, the 'Series' column gives a name for the dataset to distinguish it from other data sets of the same population.}
\item{Method}{Method for the spawner count. Survey refers to a spawning ground survey; the count itself is an expansion. See the SPS (database link in sources below) for the details for the data. LadderCount is a weir or dam count. Model  refers to the DABOM model that expands from PIT tag data. GSI is an expandsion based on the BOXCAR model, which uses GSI to allocate fish sampled at Lower Granite Dam.}
\item{Citation}{Data citation.}
\item{Contributor}{Data contributor.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{ESU}{ESU or DPS name.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{COMMON_POPULATION_NAME}{Population name.}
\item{SPECIES}{Species name.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{BROOD_YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
}
}
\source{
Oregon Department of Fish and Wildlife
}
\description{
Spawner and fraction wild data. Species: Chinook salmon. Method:

Spawner and fraction wild data. Species: chinook.
\itemize{
\item Clackamas River:
}

Spawner and fraction wild data. Species: Chinook salmon.
\itemize{
\item Willamette Falls ESC: LadderCount
}

Spawner and fraction wild data. Species: chinook.
\itemize{
\item North Santiam River:
}

Spawner and fraction wild data. Species: Chinook salmon.
\itemize{
\item South Santiam River:
}

Spawner and fraction wild data. Species: chinook.
\itemize{
\item McKenzie River:
}

Spawner and fraction wild data. Species: Chinook salmon.
\itemize{
\item Middle Fork Willamette River:
}
}
\examples{
data('Salmon Chinook (Upper Willamette River ESU)-UWchinook2020')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=NUMBER_OF_SPAWNERS)) + 
  geom_point(na.rm = TRUE) +
  ggtitle('Salmon Chinook (Upper Willamette River ESU)-UWchinook2020') +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=FRACWILD)) +
  geom_point(na.rm = TRUE) +
  ggtitle('Fraction Wild') +
  ylim(0,1) +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))
}
\references{
Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}

Original source: Citations with the format CAXMMDDYYYY indicate that the data are from the Coordinated Assessments Partnership (CAP) database and were downloaded on on MM-DD-YYYY. \url{https://www.streamnet.org/home/data-maps/fish-hlis/}.
\itemize{
\item Clackamas River: CAX05262020
\item Willamette Falls ESC: pers communication Jamie Anthony
\item North Santiam River: CAX05262020
\item South Santiam River: CAX05262020
\item McKenzie River: CAX05262020
\item Middle Fork Willamette River: CAX05262020
}
}
