% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Salmon chum (Columbia River ESU).R
\docType{data}
\name{Salmon chum (Columbia River ESU)}
\alias{Salmon chum (Columbia River ESU)}
\title{Salmon chum (Columbia River ESU)}
\format{
A csv file.
\describe{
\item{Citation}{Data citation.}
\item{Contributor}{Data contributor.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{ESU}{ESU or DPS name.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{COMMON_POPULATION_NAME}{Population name.}
\item{SPECIES}{Species name.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
}
}
\source{
Washington Department of Fish and Wildlife
}
\description{
Spawner and fraction wild data. Species: Chum salmon. Method: Survey
}
\examples{
data('Salmon chum (Columbia River ESU)')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=YEAR, y=NUMBER_OF_SPAWNERS)) + geom_point(na.rm = TRUE) +
  ggtitle('Salmon chum (Columbia River ESU)') +
  facet_wrap(~COMMON_POPULATION_NAME)

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=YEAR, y=FRACWILD)) +
 geom_point(na.rm = TRUE) +
 ggtitle('Fraction Wild') +
 ylim(0,1) +
 facet_wrap(~COMMON_POPULATION_NAME)
}
\references{
Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}

Original source: CAX05262020
}
