% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Salmon-coho-Oregon-Coast-ESU-OCcoho2020.R
\docType{data}
\name{Salmon-coho-Oregon-Coast-ESU-OCcoho2020}
\alias{Salmon-coho-Oregon-Coast-ESU-OCcoho2020}
\title{Salmon-coho-Oregon-Coast-ESU-OCcoho2020}
\format{
A csv file.
\describe{
\item{SERIES}{For certain populations, there are multiple types of spawner count data. For example survey and model counts that are based on different methods. In this case, the 'Series' column gives a name for the dataset to distinguish it from other data sets of the same population.}
\item{ESU}{ESU or DPS name.}
\item{SPECIES}{Species name.}
\item{METHOD}{Method for the spawner count. Survey refers to a spawning ground survey; the count itself is an expansion. See the SPS (database link in sources below) for the details for the data. LadderCount is a weir or dam count. Model  refers to the DABOM model that expands from PIT tag data. GSI is an expansion based on the BOXCAR model, which uses GSI to allocate fish sampled at Lower Granite Dam.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{POPULATION_NAME}{Long population name.}
\item{COMMON_POPULATION_NAME}{Shorter population name used in Viability Report figures and tables.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
\item{CONTRIBUTOR}{Data contributor. Listed under Source below.}
\item{COMPILER}{NWFSC staff who compiled the data file: Mari Williams}
\item{CITATION}{Data citation. Listed under References below.}
}
}
\source{
Oregon Department of Fish and Wildlife
}
\description{
Spawner and fraction wild data. Species: Coho salmon. Method: Survey. \href{../data/Salmon-coho-Oregon-Coast-ESU-OCcoho2020.html}{View raw data} or \href{../data/Salmon-coho-Oregon-Coast-ESU-OCcoho2020.csv}{Download raw data}
}
\examples{
data('Salmon-coho-Oregon-Coast-ESU-OCcoho2020')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=YEAR, y=NUMBER_OF_SPAWNERS)) + 
  geom_point(na.rm = TRUE) +
  ggtitle('Salmon-coho-Oregon-Coast-ESU-OCcoho2020') +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), 
   strip.text.x = element_text(size = 6))

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=YEAR, y=FRACWILD)) +
  geom_point(na.rm = TRUE) +
  ggtitle('Fraction Wild') +
  ylim(0,1) +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), 
   strip.text.x = element_text(size = 6))
}
\references{
Ford, M.J., et al. 2022. Biological Viability Assessment Update for Pacific Salmon and Steelhead Listed Under the Endangered Species Act: Pacific Northwest. U.S. Department of Commerce, NOAA Technical Memorandum NMFS-NWFSC-171. https://doi.org/10.25923/kq2n-ke70

Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}

Original source:
\itemize{
\item Alsea River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Beaver Creek: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Coos Bay: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Coquille River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Floras Creek: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Lower Umpqua: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Middle Umpqua: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Necanicum River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Nehalem River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Nestucca River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item North Umpqua: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Salmon River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Siletz River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Siltcoos River (Lake): Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Siuslaw River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Sixes River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item South Umpqua: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Tahkenitch Lake: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Tenmile Lake: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Tillamook Bay: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Yaquina River: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (1994-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Oregon Mid-Coast Coho dependent population group: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (2007-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Oregon North Coast Coho dependent population group: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (2007-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
\item Oregon Mid-South Coast Coho dependent population group: Oregon Coast Coho Natural Origin Spawner Abundance Dataset (2010-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/OAI.
}
}
