% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Salmon sockeye (Ozette Lake ESU).R
\docType{data}
\name{Salmon sockeye (Ozette Lake ESU)}
\alias{Salmon sockeye (Ozette Lake ESU)}
\title{Salmon sockeye (Ozette Lake ESU)}
\format{
A csv file.
\describe{
\item{ESU}{ESU or DPS name.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{COMMON_POPULATION_NAME}{Population name.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{SPECIES}{Species name.}
\item{BROOD_YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
}
}
\description{
Spawner and fraction wild data. Species: Sockeye.
}
\examples{
data('Salmon sockeye (Ozette Lake ESU)')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=NUMBER_OF_SPAWNERS)) + 
  geom_point(na.rm = TRUE) +
  ggtitle('Salmon sockeye (Ozette Lake ESU)') +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=FRACWILD)) +
  geom_point(na.rm = TRUE) +
  ggtitle('Fraction Wild') +
  ylim(0,1) +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))
}
\references{
Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}
}
