% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steelhead (Lower Columbia River
%   DPS)-LCsthd2020.R
\docType{data}
\name{Steelhead (Lower Columbia River DPS)-LCsthd2020}
\alias{Steelhead (Lower Columbia River DPS)-LCsthd2020}
\title{Steelhead (Lower Columbia River DPS)-LCsthd2020}
\format{
A csv file.
\describe{
\item{SERIES}{For certain populations, there are multiple types of spawner count data. For example survey and model counts that are based on different methods. In this case, the 'Series' column gives a name for the dataset to distinguish it from other data sets of the same population.}
\item{ESU}{ESU or DPS name.}
\item{SPECIES}{Species name.}
\item{METHOD}{Method for the spawner count. Survey refers to a spawning ground survey; the count itself is an expansion. See the SPS (database link in sources below) for the details for the data. LadderCount is a weir or dam count. Model  refers to the DABOM model that expands from PIT tag data. GSI is an expansion based on the BOXCAR model, which uses GSI to allocate fish sampled at Lower Granite Dam.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{POPULATION_NAME}{Long population name.}
\item{COMMON_POPULATION_NAME}{Shorter population name used in Viability Report figures and tables.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
\item{CONTRIBUTOR}{Data contributor. Listed under Source below.}
\item{COMPILER}{NWFSC staff who compiled the data file: Mari Williams}
\item{CITATION}{Data citation. Listed under References below.}
}
}
\source{
\itemize{
\item Coweeman Winter Steelhead: Washington Department of Fish and Wildlife
\item North Fork Toutle Winter Steelhead: Washington Department of Fish and Wildlife
\item South Fork Toutle Winter Steelhead: Washington Department of Fish and Wildlife
\item Upper Cowlitz and Cispus Winter Steelhead: Washington Department of Fish and Wildlife
\item Tilton Winter Steelhead: Washington Department of Fish and Wildlife
\item Kalama Winter Steelhead: Washington Department of Fish and Wildlife
\item East Fork Lewis Winter Steelhead: Washington Department of Fish and Wildlife
\item Clackamas River: Oregon Department of Fish and Wildlife
\item Sandy River: Oregon Department of Fish and Wildlife
\item Washougal Winter Steelhead: Washington Department of Fish and Wildlife
\item Upper Gorge (Columbia) Winter Steelhead: Washington Department of Fish and Wildlife
\item Hood River: Oregon Department of Fish and Wildlife
\item Kalama Summer Steelhead: Washington Department of Fish and Wildlife
\item East Fork Lewis Summer Steelhead: Washington Department of Fish and Wildlife
\item Washougal Summer Steelhead: Washington Department of Fish and Wildlife
\item Wind River (Upper Gorge) Summer Steelhead: Washington Department of Fish and Wildlife
}
}
\description{
Spawner and fraction wild data. Species: Steelhead. Method: Survey. \href{../data/Steelhead (Lower Columbia River DPS)-LCsthd2020.html}{View raw data} or \href{../data/Steelhead (Lower Columbia River DPS)-LCsthd2020.csv}{Download raw data}
}
\examples{
data('Steelhead (Lower Columbia River DPS)-LCsthd2020')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=YEAR, y=NUMBER_OF_SPAWNERS)) + 
  geom_point(na.rm = TRUE) +
  ggtitle('Steelhead (Lower Columbia River DPS)-LCsthd2020') +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), strip.text.x = element_text(size = 6))

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=YEAR, y=FRACWILD)) +
  geom_point(na.rm = TRUE) +
  ggtitle('Fraction Wild') +
  ylim(0,1) +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), strip.text.x = element_text(size = 6))
}
\references{
Ford, M.J., et al. 2022. Biological Viability Assessment Update for Pacific Salmon and Steelhead Listed Under the Endangered Species Act: Pacific Northwest. U.S. Department of Commerce, NOAA Technical Memorandum NMFS-NWFSC-171. https://doi.org/10.25923/kq2n-ke70

Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}

Original source:
\itemize{
\item Coweeman Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1987-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item North Fork Toutle Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (2005-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item South Fork Toutle Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1981-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Upper Cowlitz and Cispus Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1997-2018). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Tilton Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1999-2018). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Kalama Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1977-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item East Fork Lewis Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1986-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Clackamas River: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1974-2019). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Sandy River: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (2010-2019). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Washougal Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1991-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Upper Gorge (Columbia) Winter Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (2000-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Hood River: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (2011-2018). Spawner abundance data. Oregon Department of Fish and Wildlife. Protocol and Methods available at http://odfwrecoverytracker.org/metadata/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Kalama Summer Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1977-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item East Fork Lewis Summer Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1996-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Washougal Summer Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1953-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
\item Wind River (Upper Gorge) Summer Steelhead: Lower Columbia River Steelhead Natural Origin Spawner Abundance Dataset (1989-2019). Spawner abundance data. Washington Department of Fish and Wildlife. Protocol and Methods available at https://fortress.wa.gov/dfw/score/score/. Accessed from www.cax.streamnet.org vers May 26 2020 10:00PM by Mari Williams, NOAAF NWFSC/PSMFC.
}
}
