% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steelhead (Middle Columbia River
%   DPS)-ICMC2021.R
\docType{data}
\name{Steelhead (Middle Columbia River DPS)-ICMC2021}
\alias{Steelhead (Middle Columbia River DPS)-ICMC2021}
\title{Steelhead (Middle Columbia River DPS)-ICMC2021}
\format{
A csv file.
\describe{
\item{Series}{For certain populations, there are multiple types of spawner count data. For example survey and model counts that are based on different methods. In this case, the 'Series' column gives a name for the dataset to distinguish it from other data sets of the same population.}
\item{Method}{Method for the spawner count. Survey refers to a spawning ground survey of some type. See the SPS (database link in sources below) for the details for the data. LadderCount is a weir or dam count.}
\item{Citation}{Data citation.}
\item{Contributor}{Data contributor.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{ESU}{ESU or DPS name.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{COMMON_POPULATION_NAME}{Population name.}
\item{SPECIES}{Species name.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{BROOD_YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
}
}
\source{
\itemize{
\item Klickitat River: Yakama Nation Fisheries
\item Fifteenmile Creek: Yakama Nation Fisheries
\item Deschutes River Westside: Oregon Department of Fish and Wildlife
\item Deschutes River Eastside: Oregon Department of Fish and Wildlife
\item Rock Creek: Yakama Nation Fisheries
\item John Day River Lower Mainstem Tributaries: Oregon Department of Fish and Wildlife
\item North Fork John Day River: Oregon Department of Fish and Wildlife
\item Middle Fork John Day River: Oregon Department of Fish and Wildlife
\item South Fork John Day River: Oregon Department of Fish and Wildlife
\item John Day River Upper Mainstem: Oregon Department of Fish and Wildlife
\item Satus Creek: Yakama Nation Fisheries
\item Toppenish Creek: Yakama Nation Fisheries
\item Naches River: Yakama Nation Fisheries
\item Yakima River upper mainstem: Yakama Nation Fisheries
\item Umatilla River: Oregon Department of Fish and Wildlife
\item Walla Walla River: Oregon Department of Fish and Wildlife
\item Touchet River: Washington Department of Fish and Wildlife
}
}
\description{
Spawner and fraction wild data. Species: Steelhead. Method: Survey
}
\examples{
data('Steelhead (Middle Columbia River DPS)-ICMC2021')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=NUMBER_OF_SPAWNERS)) + 
  geom_point(na.rm = TRUE) +
  ggtitle('Steelhead (Middle Columbia River DPS)-ICMC2021') +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), strip.text.x = element_text(size = 6))

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=BROOD_YEAR, y=FRACWILD)) +
  geom_point(na.rm = TRUE) +
  ggtitle('Fraction Wild') +
  ylim(0,1) +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), strip.text.x = element_text(size = 6))
}
\references{
Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}

Original source: Citations with the format CAXMMDDYYYY indicate that the data are from the Coordinated Assessments Partnership (CAP) database and were downloaded on on MM-DD-YYYY. \url{https://www.streamnet.org/home/data-maps/fish-hlis/}.
\itemize{
\item Klickitat River: CAX05262020
\item Fifteenmile Creek: CAX05262020
\item Deschutes River Westside: CAX05262020
\item Deschutes River Eastside: CAX05262020
\item Rock Creek: CAX05262020
\item John Day River Lower Mainstem Tributaries: CAX05262020
\item North Fork John Day River: CAX05262020
\item Middle Fork John Day River: CAX05262020
\item South Fork John Day River: CAX05262020
\item John Day River Upper Mainstem: CAX05262020
\item Satus Creek: CAX07012020
\item Satus Creek: CAX07012020 5yr avg FracWild 2017 & 2018
\item Toppenish Creek: CAX07012020
\item Toppenish Creek: CAX07012020 5yr avg FracWild 2017 & 2018
\item Naches River: CAX07012020
\item Naches River: CAX07012020 5yr avg FracWild 2017 & 2018
\item Yakima River upper mainstem: CAX07012020
\item Yakima River upper mainstem: CAX07012020 5yr avg FracWild 2017 & 2018
\item Umatilla River: CAX05262020
\item Walla Walla River: CAX05262020
\item Touchet River: CAX05262020
}
}
