% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steelhead (Puget Sound DPS)-PSsteelhead2020.R
\docType{data}
\name{Steelhead (Puget Sound DPS)-PSsteelhead2020}
\alias{Steelhead (Puget Sound DPS)-PSsteelhead2020}
\title{Steelhead (Puget Sound DPS)-PSsteelhead2020}
\format{
A csv file.
\describe{
\item{SERIES}{For certain populations, there are multiple types of spawner count data. For example survey and model counts that are based on different methods. In this case, the 'Series' column gives a name for the dataset to distinguish it from other data sets of the same population.}
\item{ESU}{ESU or DPS name.}
\item{SPECIES}{Species name.}
\item{METHOD}{Method for the spawner count. Survey refers to a spawning ground survey; the count itself is an expansion. See the SPS (database link in sources below) for the details for the data. LadderCount is a weir or dam count. Model  refers to the DABOM model that expands from PIT tag data. GSI is an expansion based on the BOXCAR model, which uses GSI to allocate fish sampled at Lower Granite Dam.}
\item{MAJOR_POPULATION_GROUP}{Major population group within the ESU or DPS.}
\item{NMFS_POPID}{NWFSC id for the population.}
\item{POPULATION_NAME}{Long population name.}
\item{COMMON_POPULATION_NAME}{Shorter population name used in Viability Report figures and tables.}
\item{RUN_TIMING}{Run timing (fall, winter, spring, etc). If missing, see the ESU/DPS name or may not be applicable for the species.}
\item{YEAR}{Integer in XXXX format. The year that the count was made. See details for ESU or DPS specific comments.}
\item{NUMBER_OF_SPAWNERS}{Total spawners or escapement count. -99 means missing count. May be integer or decimal.}
\item{FRACWILD}{Decimal between 0 and 1. Fraction wild or natural. -99 means missing information.}
\item{CONTRIBUTOR}{Data contributor. Listed under Source below.}
\item{COMPILER}{NWFSC staff who compiled the data file: Katie Barnes}
\item{CITATION}{Data citation. Listed under References below.}
}
}
\source{
Washington Department of Fish and Wildlife
}
\description{
Spawner and fraction wild data. Species: Steelhead. Method: Survey. \href{../data/Steelhead (Puget Sound DPS)-PSsteelhead2020.html}{View raw data} or \href{../data/Steelhead (Puget Sound DPS)-PSsteelhead2020.csv}{Download raw data}
}
\examples{
data('Steelhead (Puget Sound DPS)-PSsteelhead2020')
library(ggplot2)
out$NUMBER_OF_SPAWNERS[out$NUMBER_OF_SPAWNERS==-99] <- NA
ggplot(out, aes(x=YEAR, y=NUMBER_OF_SPAWNERS)) + 
  geom_point(na.rm = TRUE) +
  ggtitle('Steelhead (Puget Sound DPS)-PSsteelhead2020') +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), strip.text.x = element_text(size = 6))

out$FRACWILD[out$FRACWILD == -99] <- NA
ggplot(out, aes(x=YEAR, y=FRACWILD)) +
  geom_point(na.rm = TRUE) +
  ggtitle('Fraction Wild') +
  ylim(0,1) +
  facet_wrap(~COMMON_POPULATION_NAME) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), strip.text.x = element_text(size = 6))
}
\references{
Ford, M.J., editor. 2022. Biological Viability Assessment Update for Pacific Salmon and Steelhead Listed Under the Endangered Species Act: Pacific Northwest. U.S. Department of Commerce, NOAA Technical Memorandum NMFS-NWFSC-171. https://doi.org/10.25923/kq2n-ke70

Salmon Population Summaries (SPS) Database: \url{https://www.fisheries.noaa.gov/resource/tool-app/salmon-population-summaries-sps}

Original source:
\itemize{
\item Samish River and Bellingham Bay Tributaries - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1979-2018). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Nooksack River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (2004-2018). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Skagit River - summer and winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1978-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Stillaguamish River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1985-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Snohomish/Skykomish Rivers - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1981-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Pilchuck River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1981-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Tolt River - summer Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1985-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Snoqualmie River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1981-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item North Lake Washington and Lake Sammamish - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1984-1999). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Cedar River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1984-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Green River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1978-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item White River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1983-2018). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Puyallup River/Carbon River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1992-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Nisqually River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1980-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item South Hood Canal Tributaries - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1998-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item East Hood Canal Tributaries - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1981-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Skokomish River - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1982-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item West Hood Canal Tributaries - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1995-2018). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Dungeness River - summer and winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1988-2019). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Strait of Juan de Fuca Tributaries - winter Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (1984-2018). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
\item Elhwa Steelhead: Puget Sound Steelhead Natural Origin Spawner Abundance Dataset (2013-2018). Spawner abundance Data. Washington Department of Fish and Wildlife. Protocols and Methods available at https://fortress.wa.gov/dfw/score/score/species/steelhead.jsp?species=Steelhead. Accessed from WA SCORE on March 9, 2021 by Katie Barnas, NOAAF NWFSC.
}
}
