% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrdata_names.R
\docType{data}
\name{vrdata_names}
\alias{vrdata_names}
\title{VRData names}
\format{
A vector of names.
#' \describe{
\item{esu_dps}{Allowed ESU or DPS names.}
\item(species}{Allowed species names.}
\item(run_timing}{Allowed run timings. May be blank.}
}
}
\description{
Allowed values for data columns.
#'
}
\examples{
data('vrdata_names')
cat(c("ESU and DPS", esu_dps), sep="\n")
cat(c("\n", "Species", species), sep="\n")
cat(c("\n", "Run Timing", run_timing), sep="\n")
}
\author{
Eli Holmes (NOAA)
}
