% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_raw.R
\name{process_data_raw}
\alias{process_data_raw}
\title{Process data-raw}
\usage{
process_data_raw()
}
\description{
Process's the Roxygen header on the csv files in data-raw into R files for each data set into
data folder and R folder
}
\keyword{internal}
